✨ feat : websocket jwt 연결

+++ src/main/java/study/withkbo/chat/service/ChatRoomService.java
@@ -1,14 +1,19 @@
 package study.withkbo.chat.service;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import study.withkbo.chat.dto.request.ChatRoomRequestDto;
 import study.withkbo.chat.entity.ChatInvitation;
 import study.withkbo.chat.entity.ChatMessage;
+import study.withkbo.chat.entity.ChatParticipants;
 import study.withkbo.chat.entity.ChatRoom;
 import study.withkbo.chat.repository.ChatInvitationRepository;
 import study.withkbo.chat.repository.ChatMessageRepository;
+import study.withkbo.chat.repository.ChatParticipantsRepository;
 import study.withkbo.chat.repository.ChatRoomRepository;
 import study.withkbo.exception.common.CommonError;
 import study.withkbo.exception.common.CommonException;
@@ -17,6 +22,7 @@ import study.withkbo.user.repository.UserRepository;
 import study.withkbo.user.service.UserService;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -25,6 +31,8 @@ public class ChatRoomService {
     private final ChatRoomRepository chatRoomRepository;
     private final SimpMessagingTemplate template;
     private final ChatInvitationRepository chatInvitationRepository;
+    private final UserRepository userRepository;
+    private final ChatParticipantsRepository chatParticipantsRepository;
     private final ChatMessageRepository chatMessageRepository;
 
     // 채팅방 생성
@@ -47,15 +55,23 @@ public class ChatRoomService {
         chatInvitation.setStatus("생성됨");
         chatInvitationRepository.save(chatInvitation);
 
+        ChatParticipants chatParticipants = new ChatParticipants();
+        chatParticipants.setRoom(saveRoom);
+        chatParticipants.setUser(inviter);
+        chatParticipantsRepository.save(chatParticipants);
+
         return saveRoom;
     }
 
     // 전체 채팅방 조회
-    public List<ChatRoom> getChatRoomByUserId(Long userId) {
+    public List<ChatRoomRequestDto> getChatRoomByUserId(Long userId) {
         if (userId == null) {
             throw new CommonException(CommonError.INVALID_INPUT);
         }
-        return chatRoomRepository.findAllByUserId(userId);
+        List<ChatRoom> chatRooms = chatRoomRepository.findChatRoomByUserId(userId);
+        return chatRooms.stream()
+                .map(room -> new ChatRoomRequestDto(room.getId(), room.getRoomName(), room.getCreatedDate()))
+                .collect(Collectors.toList());
     }
 
     // 특정 채팅방 조회(부분 문자열 가능)
@@ -64,15 +80,19 @@ public class ChatRoomService {
     }
 
     // 채팅방 나가기
+    @Transactional
     public void leaveChatRoom(Long roomId, @AuthenticationPrincipal User user) {
-        getOutChatRoom(roomId, user);
+         getOutChatRoom(roomId, user);
         checkAndDeleteChatRoom(roomId);
     }
 
     // 채팅방에서 유저 삭제
+    @Transactional
     public void getOutChatRoom(Long roomId, @AuthenticationPrincipal User user) {
-        ChatRoom chatRoom = chatRoomRepository.findById(roomId)
-                .orElseThrow(() -> new CommonException(CommonError.NOT_FOUND));
+        ChatRoom chatRoom = chatRoomRepository.findByIdForUpdate(roomId);
+        if (chatRoom == null) {
+            throw new CommonException(CommonError.NOT_FOUND);
+        }
         List<ChatInvitation> invitationsFromUser= chatInvitationRepository.findByRoomAndInviter(chatRoom, user);
         List<ChatInvitation> invitationsToUser= chatInvitationRepository.findByRoomAndInviteeContaining(chatRoom, user);
         // List 합침
@@ -81,12 +101,13 @@ public class ChatRoomService {
             throw new CommonException(CommonError.NOT_FOUND);
         }
         chatInvitationRepository.deleteByRoom(chatRoom);
+
     }
 
     // 채팅방이 0명이 될 경우 채팅방 삭제
-    private void checkAndDeleteChatRoom(Long roomId) {
-        ChatRoom chatRoom = chatRoomRepository.findById(roomId)
-                .orElseThrow(() -> new CommonException(CommonError.NOT_FOUND));
+    @Transactional
+    public void checkAndDeleteChatRoom(Long roomId) {
+        ChatRoom chatRoom = chatRoomRepository.findByIdForUpdate(roomId);
         long userCount = chatInvitationRepository.countByRoom(chatRoom);
         if (userCount == 0) {
             chatRoomRepository.delete(chatRoom);
@@ -94,9 +115,12 @@ public class ChatRoomService {
     }
 
     // 채팅방 삭제
+    @Transactional
     public void ridOffChatRoom(Long roomId) {
-        ChatRoom chatRoom = chatRoomRepository.findById(roomId)
-                .orElseThrow(() -> new CommonException(CommonError.NOT_FOUND));
+        ChatRoom chatRoom = chatRoomRepository.findByIdForUpdate(roomId);
+        if (chatRoom == null) {
+            throw new CommonException(CommonError.NOT_FOUND);
+        }
         chatInvitationRepository.deleteByRoom(chatRoom);
         chatMessageRepository.deleteByRoom(chatRoom);
         chatRoomRepository.delete(chatRoom);
